requirements.txt
openai
requests
beautifulsoup4
lxml
rich
python-dotenv
typer[all]
httpx
fake-useragent
deepl
questionary
selenium
PyPDF2
pandas